Clear __init__.py, move ProtonDrive to driver.
